Implementation of authorisation

+++ App/controllers/ListingController.php
@@ -2,6 +2,8 @@
     namespace App\Controllers;
     use Framework\Database;
     use Framework\Validation;
+    use Framework\Session;
+    use Framework\Authorisation;
     class ListingController {
         protected $db;
 
@@ -60,7 +62,7 @@
 
             $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-            $newListingData['user_id'] = 1;
+            $newListingData['user_id'] = Session::get('user')['id'];
             $newListingData = array_map('sanitise', $newListingData);
 
             $requiredFields = ['title', 'description', 'salary', 'email', 'city', 'state'];
@@ -97,6 +99,7 @@
 
                 $query = "INSERT INTO listings ({$fields}) VALUES ({$values})";
                 $this->db->query($query, $newListingData);
+                Session::setFlashMessage('success_message', 'Listing created successfully');
                 redirect('/listings');
             }
         }
@@ -115,13 +118,20 @@
 
             $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+            // Check if listing exists
             if(!$listing) {
                 ErrorController::notFound('Listing not found');
                 return;
             }
+
+            // Authorisation
+            if(!Authorisation::isOwner($listing->user_id)) {
+                Session::setFlashMessage('error_message', 'Your are not authorised to delete this listing');
+                return redirect('/listings/' . $listing->id);
+            }
             $this->db->query('DELETE FROM listings WHERE id = :id', $params);
             // Set flash message
-            $_SESSION['success_message'] = 'Listing deleted successfully';
+            Session::setFlashMessage('success_message', 'Listing deleted successfully');
             redirect('/listings');
         }
 
@@ -142,6 +152,12 @@
             if (!$listing) {
                 ErrorController::notFound('Listing not found');
                 return;
+            }
+
+            // Authorisation
+            if(!Authorisation::isOwner($listing->user_id)) {
+                Session::setFlashMessage('error_message', 'Your are not authorised to update this listing');
+                return redirect('/listings/' . $listing->id);
             }
 
             loadView('listings/edit', [
@@ -167,6 +183,13 @@
                 ErrorController::notFound('Listing not found');
                 return;
             }
+
+            // Authorisation
+            if(!Authorisation::isOwner($listing->user_id)) {
+                Session::setFlashMessage('error_message', 'Your are not authorised to update this listing');
+                return redirect('/listings/' . $listing->id);
+            }
+
             $allowedFields = ['title', 'description', 'salary', 'tags', 'company', 'address', 'city', 'state', 'phone', 'email', 'requirements', 'benefits'];
             $updateValues = [];
             $updateValues = array_intersect_key($_POST, array_flip($allowedFields));
@@ -196,7 +219,7 @@
                 $updateQuery = "UPDATE listings SET {$updateFields} WHERE id = :id";
                 $updateValues['id'] = $id;
                 $this->db->query($updateQuery, $updateValues);
-                $_SESSION['success_message'] = 'Listing Updated';
+                Session::setFlashMessage('success_message', 'Listing updated successfully');
                 redirect('/listings/' . $id);
             }
         }
